Refactored credential contexts
