+++ README.md
@@ -1,7 +1,8 @@
 # IOT
 Public library of example and usable code for integrating an IOT project :)
 
-Python Branch: [![Upload Python Package](https://github.com/ActuallyHappening/IOT/actions/workflows/python.yml/badge.svg)](https://github.com/ActuallyHappening/IOT/actions/workflows/python.yml)
+Python Branch: [![Test, Build and upload :)](https://github.com/ActuallyHappening/IOT/actions/workflows/python.yml/badge.svg)](https://github.com/ActuallyHappening/IOT/actions/workflows/python.yml)
+
 React Branch: [![Netlify Status](https://api.netlify.com/api/v1/badges/ca620688-dbf5-429c-94e8-23b819edf36c/deploy-status)](https://app.netlify.com/sites/aidentaldetect/deploys)
 
 # Usage for Thermal Sensor
